+++ src/core/transport/chttp2/timeout_encoding.c
@@ -126,7 +126,7 @@ void grpc_chttp2_encode_timeout(gpr_timespec timeout, char *buffer) {
                (int)(timeout.tv_sec * 1000000) +
                    (timeout.tv_nsec / 1000 + (timeout.tv_nsec % 1000 != 0)));
   } else {
-    enc_seconds(buffer, timeout.tv_sec + (timeout.tv_nsec != 0));
+    enc_seconds(buffer, (long)(timeout.tv_sec) + (timeout.tv_nsec != 0));
   }
 }
